+++ rmp/tests/func/encode/null.rs
@@ -12,9 +12,5 @@ fn pass_pack() {
 #[test]
 fn fail_pack_too_small_buffer() {
     let mut buf = [];
-
-    match write_nil(&mut &mut buf[..]) {
-        Err(FixedValueWriteError(..)) => (),
-        other => panic!("unexpected result: {:?}", other)
-    }
+    write_nil(&mut &mut buf[..]).err().unwrap();
 }